Escape order by and group by columns
Special column names, like status, can also cause SphinxQL to raise an
error if not escaped with backticks.  All columns are now escaped,
except those that are sphinx expressions, e.g. @weight.  In order to
handle escaping order_by columns, which can come in with a direction
such as "created_at DESC", the column had to be split on spaces and only
the first part escaped.